requirements.txt
langchain
langchain-core
langchain-community
langchain-chroma
chromadb
langchain-google-genai
python-dotenv
